Allow control of distResisted, prevent DOM deletion on pull, add releasing callback:
This commit implements a small series of changes:

1. The current `resistanceFunction` only allows you to implement drags that
begin as difficult pulls but become easier as you pull. That's because the
value between 0-1 that you return is multiplied (basically) by how far the user
has pulled to arrive at the `distResisted` to determine how far to move the PTR
elements down on screen. This change replaces `resistanceFunction` with
`distanceResistedFunction`, which takes in `distExtra` (how much you have
dragged while accounting for the `distIgnore` param) and `distThreshold` (the
distance we should arrive at before triggering refresh) to return the
`distResisted` value (aka how far the PTR should be considered dragged).
1. The `setTimeout` that cleans up the PTR element in the DOM makes the library
fairly "glitchy" for me. It reuses the `refreshTimeout` to wait until deleting
the PTR elements from the DOM. Whenever I modify any of the `init` params
outside of their default values, then I frequently encounter the problem where
the PTR element is deleted while I'm PTR-ing for a second time because the
timeout fires during my second drag. This makes the element disappear mid-pull
and I have to start over again. This change moves the element cleanup to the
handler destruction call.
1. When transitioning from pulling to releasing, the consumer might want
something to happen, so I added a callback `onReleasingWhilePulling` that is
called at this time.